Refactoring Startup and personalize message
Creating a new extension for refactoring startup and personalize messages in postValidators withMessage.
